+++ docs/changelog.rst
@@ -6,7 +6,7 @@ v0.2 (Not yet released)
 
 * aiger: ``on_bad_state``, ``on_constraint``, ``on_fairness``, ``on_justice``, ``on_justice_header``, ``on_justice_name``.
 * aiger: ``read_ascii_aiger``.
-* pla: ``on_keyword``.
+* pla: ``on_keyword``, ``pla_writer``.
 
 v0.1 (April 27, 2018)
 ---------------------